+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Add
+
+- Option to delay the start of processes to space them out.
+
 ## 1.0.0 - 2018-09-10
 
 - Initial version of the process manager.